+++ timerpopup.py
@@ -37,7 +37,7 @@ class TimerDialog(QDialog):
 
         # Force gui to display correctly
         self.show()
-        self.move(parent.frameGeometry().center().x() - self.frameGeometry().width() / 2, parent.frameGeometry().center().y() - self.frameGeometry().height() / 2)
+        self.move(int(parent.frameGeometry().center().x() - self.frameGeometry().width() / 2), int(parent.frameGeometry().center().y() - self.frameGeometry().height() / 2))
         # Setup mouse event handling
         self.oldPos = self.pos()
